Problem set 1: Credit
requirements:
Get user input of credit card number
Do a checksum on the number using Luhn’s Algorithm
Verify the number is the correct number of digits
Verify the number starts with the correct number for a valid AMEX, MC, or VISA
For this exercise the following were true:
AMEX: 15 digits, starts with 34 or 37
MC: 16 digits, starts with 51, 52, 53, 54, 55
VISA: 13 or 16 digits and starts with 4
Correct program had to output one of the following "AMEX", "MASTERCARD", "VISA", or "INVALID"

There were a couple places I probably could have used a function, the modulo operations on the CC number for instance. At the time I didn't think the complexity of the program and it's length called for a function. Once I'm more experienced perhaps that would change.