Lab 2: scrabble
Ask for two words. Then call function with each word to determine a scrabble score for each word. Then output whether player1 wins, player2 wins, or it's a tie.
Part of this code was supplied by Lab, the two sections with comments starting with "TODO" are the sections created by me.