problem set 2: substitution
pset requirements:
- Program accepts 1 argument, the key, and is not case sensitive
- If either 0 or 2+ arguments print a message reminding user of the usage of the program "Usage: ./substitution key" and return 1 from main
- If argument is invalid, such as the key is not 26 characters and or the key does not list each letter exactly 1 time, or has non alpha characters it should print a message and return 1 from main
- Your program must output ciphertext: (without a newline) followed by the plaintext’s corresponding ciphertext, with each alphabetical character in the plaintext substituted for the corresponding character in the ciphertext; non-alphabetical characters should be outputted unchanged.
- Program must preserve the case of the user inputted plain text.
- After outputting ciphertext the program should return a new line then return 0 from main

I wasn't sure which would be better design, to have the convert_plaintext function do as much as it did or to have more smaller functions and call them from either main or as helper functions to convert_plaintext. In the end I decided to leave it mostly in convert_plaintext.
The other possible change I could have made was to use bool for a the two functions that were basically just returning 0 or 1. I didn't think of it until after I had written the functions so I left them using integers with a mental note to try to use bool if appropriate later. I figured in this case either worked about as well and I had already written the functions so didn't think a re-write would be a good return on investment as far as time usage.