pset4 filter-more original upload 3/16/22
For this problem set they provided a handful of files. Our task was to implement the functions declared in the helpers.c file creating different filter effects.
execute the file with 3 arguments with a switch included to tell the program which filter to use ./filter -g/r/b/e INFILE.bmp OUTFILE.bmp 

Grayscale
Take an input .bmp file and make it grayscale by taking the RGB values of each pixel, finding the average of the 3 values, then setting the pixel's RGB values to that average.
I managed to implement this with minimal issues. We had all the building blocks from getting tideman in the previous problem set working.

Reflect
The goal was to take an input .bmp and reverse the image horizontally and output it to a new .bmp file
This one was fairly straight forward as well. I did get caught up in one mistake though. I had created a new array to store the pixel data for 1 row, since I was swapping the pixels one row at a time. However when I initially coded it I had my double loop was iterating both i and j. This was causing a segmentation fault. I noticed running debug50 that the value for height was changing around row1 column 574 which was a bit confusing. It wasn't until I ran Valgrind and saw "“Access not within mapped region at address 0x205C693686” that I realized what I was doing. Once I fixed that it ran like a charm.

Blur
implement a blur function use a "box blur"
This was more difficult to implement as you had to setup an outer double loop to go through image[][] 2d array then inside that you needed to setup a second double loop to go through the 9 spaces around the image[i][j] pixel.  Adding to the difficulty was that for edge/corner cases you only wanted either 4 or 6 elements respectively.
At first I was using an if statement to separate the edge cases from the pixels where all 9 elements were available. I struggled trying to figure out a good way to do this. I knew I could use if statements but looking at it I knew It was going to take a bunch of them, and maybe a use of switch or two. With the number of if statements I thought I'd need I figured there had to be a better way. I was trying to find a way to have an if statement that looked for something like null if it found a element that was out of bounds of the image[][] array. I wasn't having any luck with that. I was considering asking for help in the discord for this pset as I was sure there had to be a better way of doing it and frankly I just didn't want to code like 12 if statements. When I looked at the discord the hint that was given to the last person asked if the person who'd asked the question had considered the elements of the array they were using. It made me go back and look at things again and I realized that I could setup an in if statement inside my inner double loop, that checks the 9 elements, and just rule out any that were outside the bounds of the image[][] array.
I got that straightened out and was still having issues passing check50's tests but when I was going through debug it seemed to be doing everything I was expecting it to. I could see from the output of the tests that I was returning a different value than they were expecting for the RGB values. Turns out I had misheard/remembered that the averages were of the original values. So I had to create a temp array to keep everything in then copy it out at the end to image[][] Once I did that everything worked as intended.

Edges
This one is a bit more complex than blur, you have to implement edges using a Sobel Operator. I was able to re-use a bit of code from blur so that helped. I did have to redo the meat of the double loops. It required creating a bunch of variables and 2 arrays for the Gx and Gy values. I had a bit of issue with this one, again I was returning values that weren't correct. I had to do some debugging to determine that my math'ing was correct. Once I nailed that down I knew the issue was that I wasn't getting Gx and Gy properly.  It took me a bit to get my head around the correct way to reference k and l to get the correct values out of Gx and Gy. Once I was able to get that straight things worked out.