Update credit.c original upload 2/16/22
In the file called credit.c in the credit directory, write a program that prompts the user for a credit card number and then reports (via printf) whether it is a valid American Express, MasterCard, or Visa card number, per the definitions of each’s format herein. So that we can automate some tests of your code, we ask that your program’s last line of output be AMEX\n or MASTERCARD\n or VISA\n or INVALID\n, nothing more, nothing less. For simplicity, you may assume that the user’s input will be entirely numeric (i.e., devoid of hyphens, as might be printed on an actual card) and that it won’t have leading zeroes. But do not assume that the user’s input will fit in an int! Best to use get_long from CS50’s library to get users’ input. (Why?)

Using Luhn's Algorithm determine if a CC # is valid.

I'm not sure how the tabbed indents got messed up in that one section since I only use tab to create the indents, must've bumped a backspace or something.
Removed the extra \n from line 80 as it was causing an extra carriage return which caused check50 to show it as an invalid output.