problem set 5: Speller
speller.zip is the distribution files.
The rest of the files:
staff.txt and student.txt are comparisons between my code and the staff's using the text/stoker.txt
dictionary.c is my filled in version of the file with the functions completed
Specifications:
Alright, the challenge now before you is to implement, in order, load, hash, size, check, and unload as efficiently as possible using a hash table in such a way that TIME IN load, TIME IN check, TIME IN size, and TIME IN unload are all minimized. To be sure, it’s not obvious what it even means to be minimized, inasmuch as these benchmarks will certainly vary as you feed speller different values for dictionary and for text. But therein lies the challenge, if not the fun, of this problem. This problem is your chance to design. Although we invite you to minimize space, your ultimate enemy is time. But before you dive in, some specifications from us.

-You may not alter speller.c or Makefile.
-You may alter dictionary.c (and, in fact, must in order to complete the implementations of load, hash, size, check, and unload), but you may not alter the declarations (i.e., prototypes) of load, hash, size, check, or unload. You may, though, add new functions and (local or global) variables to dictionary.c.
-You may change the value of N in dictionary.c, so that your hash table can have more buckets.
-You may alter dictionary.h, but you may not alter the declarations of load, hash, size, check, or unload.
-Your implementation of check must be case-insensitive. In other words, if foo is in dictionary, then check should return true given any capitalization thereof; none of foo, foO, fOo, fOO, fOO, Foo, FoO, FOo, and FOO should be considered misspelled.
-Capitalization aside, your implementation of check should only return true for words actually in dictionary. Beware hard-coding common words (e.g., the), lest we pass your implementation a dictionary without those same words. Moreover, the only possessives allowed are those actually in dictionary. In other words, even if foo is in dictionary, check should return false given foo's if foo's is not also in dictionary.
-You may assume that any dictionary passed to your program will be structured exactly like ours, alphabetically sorted from top to bottom with one word per line, each of which ends with \n. You may also assume that dictionary will contain at least one word, that no word will be longer than LENGTH (a constant defined in dictionary.h) characters, that no word will appear more than once, that each word will contain only lowercase alphabetical characters and possibly apostrophes, and that no word will start with an apostrophe.
-You may assume that check will only be passed words that contain (uppercase or lowercase) alphabetical characters and possibly apostrophes.
-Your spell checker may only take text and, optionally, dictionary as input. Although you might be inclined (particularly if among those more comfortable) to “pre-process” our 
 default dictionary in order to derive an “ideal hash function” for it, you may not save the output of any such pre-processing to disk in order to load it back into memory on 
 subsequent runs of your spell checker in order to gain an advantage.
-Your spell checker must not leak any memory. Be sure to check for leaks with valgrind.
-The hash function you write should ultimately be your own, not one you search for online.


my thoughts:
I continue to have issues with getting pointers correct. They make sense when I go over what they do but when I got to implement something gets lost in translation. It's still pretty common for the compiler to remind me did you mean -> instead of . 
I had a few memory errors I had to debug, I was free'ing current and temp in my check function when I hadn't malloc'd space for it, just created a veriable
my unload function also had an extra free().  Between Valgrind and Debug50 I was able to find the errors, tough not quickly by any means.
Making things worst was that check wasn't evaluating correctly, during troubleshooting I found the actual problem was my hash function was coming up with two different numbers, one when load called it and another when check did. I was able to fix that by hard coding everything in hash to use tolower(). Might be a more elegant solution out there.
I started out trying to use a recursive helper function to unload the linked lists at the end but took it out and implemented the current version while trying to track down the memory issues.