lab 6: World Cup
distribution code provided the .csv files and some of the code within tournament.py.  Our task was to implement the two parts of main indicated with the TODO and the simulate_tournament(teams) function

Program takes a filename as an argument and then outputs the percentage chance each team within has of winning the World Cup.
.csv file contains the name of the team along with it's rating.

I did fairly ok on this lab. I wasn't familiar with the csv library but some of the lecture covered reading and writing from a .csv and I was able to use that and change it to suit my needs. 

I did have a couple issues with the program at first where I was improperly handling lists and the dictionaries within but I was able to resolve them fairly quickly with debug50 once I saw what was happening line by line it made the error make sense and I was able to adjust things to pull out the part I needed incrementally.
I did have to remind myself the proper syntax for adding to a dictionary as It's been a while and I had to remember which to use between [] () and {}

The last bit I ran into was that the check50 was claiming that my function for simulate_tournament wasn't handling brackets of 4, 8, 12, and 16 properly.  It turns out I had misread the comment in the function, I was supposed to only return the team's name, not the list item which was the dictionary of {"team" : England, "rating": 1024} or whatever. I had been managing that, but it was up in main where they wanted it down in simulate_tournament so it was just a manner of moving the two lines of code down and changing the variable names to those used within the function.
I didn't actually keep track of timing on this but I finished the whole thing in about 2 hours and probably spent 30-45 mins troubleshooting the various bits. Overall I'm fairly happy with the results.