pset 6 credit
verify if a user's inputted number is a valid credit card number, AMEX, MASTERCARD, VISA, or INVALID
Use Luhn's Algorithm to determine if it is.
    Multiply every other digit by 2, starting with the number’s second-to-last digit, and then add those products’ digits together.
    Add the sum to the sum of the digits that weren’t multiplied by 2.
    If the total’s last digit is 0 (or, put more formally, if the total modulo 10 is congruent to 0), the number is valid!
Bit of bump on the cc_start function. At first I had implemented it with ints in the list and was comparing it to strings, which obviously didn't work. I was able to figure out the issue relatively quickly with debug50 and walking through the steps.

It took a bit of research to figure out the best way to iterate through the cc_string in reverse so that I could do the calculations correctly. Also I misread the part where for every other number you multiply by 2 and then add the digits to the sum, not the sum of n*2. I found that pretty quick actually as when I ran my test through debug I was able to see the sum wasn't matching the problem set's example of Luhn's Algorithm

Per the specs they wanted to re-prompt for a number if it had dashes in it, or if you put text like foo in it, but wanted invalid if it was all numbers but too short to be a valid CC. It took a bit of futzing to get it all working they way they wanted it. I was able to research a isnumeric() function which I hoped existed since I was aware of isalpha being a thing in at least some languages.

I didn't have to use main() or the helper functions, the whole app wasn't that large, but I figured it'd be good practice and maybe a bit more modular for upkeep sake (pretending for a moment this was a real app and not just some practice excercise)