pset 6: readability
The Coleman-Liau index to determine at what grade a user inputted text would be read at.
The Coleman-Liau index is computed as 0.0588 * L - 0.296 * S - 15.8, where L is the average number of letters per 100 words in the text, and S is the average number of sentences per 100 words in the text.
grade < 1 outputted as "Before Grade 1"
grade > 16: Grade 16+
every other grade just output Grade X where X is the numerical grade between 1 and 16

The hint for this problem suggested that ord and chr might be useful for this problem but I fail to see how. We're not dealing with ascii values in either direction...

I was able to use built in methods for each of the three things we needed to count. Letters: str.isalpha(), words: str.split(), sentences: str.count() with ?, !, and period for the parameters

I just realized I imported exit from sys library but never actually used it and didn't think to remove it before submitting.