pset 6: DNA
dna.zip: Has the distribution code, containing the csv files for sequences and databases used within the program
dna.py: The function longest_match() was pre-defined, our task was to implement the sections in main() marked with TODO

In a file called dna.py, implement a program that identifies to whom a sequence of DNA belongs.

The program should require as its first command-line argument the name of a CSV file containing the STR counts for a list of individuals and should require as its second command-line argument the name of a text file containing the DNA sequence to identify.
    If your program is executed with the incorrect number of command-line arguments, your program should print an error message of your choice (with print). If the correct 
    number of arguments are provided, you may assume that the first argument is indeed the filename of a valid CSV file and that the second argument is the filename of a 
    valid text file.

Your program should open the CSV file and read its contents into memory.
    You may assume that the first row of the CSV file will be the column names. The first column will be the word name and the remaining columns will be the STR sequences 
    themselves.

Your program should open the DNA sequence and read its contents into memory.

For each of the STRs (from the first line of the CSV file), your program should compute the longest run of consecutive repeats of the STR in the DNA sequence to identify. Notice that we’ve defined a helper function for you, longest_match, which will do just that!

If the STR counts match exactly with any of the individuals in the CSV file, your program should print out the name of the matching individual.
    You may assume that the STR counts will not match more than one individual.
    If the STR counts do not match exactly with any of the individuals in the CSV file, your program should print No match.

Thoughts:
I missed that I needed len() of sys.argv for determining the number of arguments. I think I was thinking of argc from C that just gives the number of arguments.
I had an issue where the line: if  int(person[STR]) == int(longest[STR]): wasn't evaluating properly. I had originally had it without casting them as int. I'm assuming one was a string and the other an int but I didn't dig deeper into it after getting it to work properly.

Overall I'm fairly happy with how quickly I finished this exercise. None of the errors I was getting were catastrophic and were relatively easy to troubleshoot.