pset 7: Fiftyville
Distribution code provided a blank log.sql file, fiftyville.db, and answers.txt. The rest are files I used to store information from queries I ran for later reference since I get pulled away fairly often and will regularly have hours/days between times to work on the problem.
Problem: CS50 duck was stolen.
What we know: took place on July 28, 2021 and took place on Humphrey Street
What we need: 
    Who the thief is,
    What city the thief escaped to, and
    Who the thief’s accomplice is who helped them escape

Thoughts:
I didn't have much trouble with Syntax on this one. SQL is fairly easy to get along with as far as we've gone. No doubt when you get into real world issues it's way harder but I'm encouraged by this brief exposure to it.
I noted in the log.sql in my final thoughts that I feel like I was really hampered by the way in which I was going about this in that I do these at work where I have the most time easily available. Unfortunately in this case I had hours between times I could work on this, and in a couple cases several days.  I didn't do a good enough job of using the log.sql to refresh my memory of where I was at.  
I also go lost in the weeds a bit in trying to prove who the accomplice was using data from the .db. Typically you'd be able to prove the accomplice bought the tickets for the thief through bank transactions and we couldn't do that here. This scenario was more simplistic in this regards and ultimately you just need to know who the thief called that day.
Lastly, I was trying to not use hardcoded information in my queries based on a general thought that hardcoding things in programs is bad and you should store it in a variable so you can maintain it easier in the future. In this particular scenario that hindered more than helped and caused me to spend a significant amount of time trying to work with sub-queries I didn't really need to when I could just go more linearly and reduce results with prior query info.