Final files for lab 8 Trivia
Index.html, trivia.js, and styles.css are the pertinent files for the lab.
Pressbutton.css - saved from a tutorial on how to get buttons to animate the way I wanted for reference
workingBits.js - file I created to keep track of parts of code that I got working correctly so I wouldn't lose track of them while troubleshooting other parts of code
jspseudocode.js - where I put my initial pseudocode for the JS part of the lab. I also was using this to keep track of things I had tried while troubleshooting code.

Part 1 I completed with little fuss and got working the same day or next day I started it.
Part 2 I tried to do first because I thought that'd be the easier of the two things to work out and then I could apply what I used there and modify it for part 1. However I beat my head against the wall off and on over the course of 3 weeks or so.  I got the part to display "correct" or "incorrect" within  the first week or so. I had some issues with it but the real issue was getting the background color of the input field to work properly. Every tutorial I found used element.style.backgroundColor to change the color but it just wouldn't work for me. I tried getComputedStyle that Stack Overflow stated was the proper way, didn't work. I tried a solution I found with element.classList.add() that also wasn't working but was supposed to be the new fangled way of doing things.
The color would only change if I clicked the submit button and nothing had been inputted into the field.
Normally I would have tried another browser earlier in this process but for whatever reason it didn't occur to me to try Edge. That's when I found it was working in Edge. But classList is supposed by Chrome... Then I found that if I typed something into the field and submitted it worked, if I select something from the Autofill it doesn't. Then I found that apparently many browsers don't let you change background color from autofill. 
Ultimately for this lab since it's a trivia thing where you wouldn't need or want to enter the same thing more than one time I found the solution of adding to the input's html a autocomplete="off" 
In other contexts a better solution would be required in order to maintain better user usability for the site but in this context I felt that it was fine.