Pset8: homepage
Whew... it's taken a really long time to do this. Some of it was that I've been busy at work and at home so didn't have as much time, some of it was procrastination, and some of it was just that I didn't feel like I knew what I was doing. I'd taken a CSS/HTML5 course through w3schools but hadn't touched on JS before and the lectures were basically just well here's a couple of the main differences, go Google the rest.  

As is noted in the planning.txt I had some pretty big ideas, or at least to me they were maybe a pro could bang them out in a day, unfortunately it being almost Sept and I have another week to go then final project I felt like I didn't have time to finish some of the things I wanted to do the way I wanted to do them. 

The API stuff for the movies page I wanted to do wasn't working very well. I wanted to get the top 10 movies of each Genre from IMDB according to their list of Genres, however the best I could do was a list of top movies over-all. Though at least I did get some API functionality working so I have a base to work off on later. In the end I decided to scrap the movies and woodworking pages for a later iteration.
I also would like to have a management page where I could add authors/books, movies, etc without having to go through and hack at the code line by line. I imagine a "addAuthor" function where it generates a UI pop-up where I can enter in the authors information including links to socials and such and it'll update the appropriate pages. I'm thinking it'll require a database of some kind and I'll have to refactor the html and add JS to make it so the HTML is generated dynamically based on some kind of For loop.
The CSS for the spacing of the hoverable side-nav is a bit of a mess and I don't care for it at all. It's something that I combined 2 or 3 tutorials to create so there may be tons of room to make it better. I feel like there has to be some kind of function I can run to dynamically figure the spacing based on how many entries and such, the biggest hurdle to my initial thought process on it was that I needed to determine how many pixels the title part was so I could have just the icons showing still. I'm sure there is a way but again it's something I ran out of time and so have pushed it to v2 of the site. The side-nav visually is pretty close to how I imagined it working which is nice, I did have to add some drop down action for smaller devices as part of the pset was that it had to be nice looking on a variety of devices and the hoverable side-nav as it was would extend down off the viewport on smaller devices at times.
There continues to be times where things I take directly from tutorials don't work. The CSS for changing the default coloring of accordions was an example. Even though there is an .accordion class it just would not change anything on the rendered page. I ended up having to drill down to the three subparts of the accordion to get it to change any of it.  I had a ton of trouble getting the footer to behave and look how I wanted. It still looks a little wonky on smaller devices but not as bad as it was at one point. 

I hope to be able to come back after I finish the final project and implement the rest of the ideas I had and tighten some things up. We'll see how long it takes to complete week 9 and the final project.  Eventually I'll update it simply because I designed it with the idea that I'd have it setup as an actual page and not just as a school project.