lab9 Birthdays
birthdays.zip is the distributed code for the lab.
app.py and index.html are the modified files with my changes. 

Just a couple minor mistakes on this one. I had originally had the month and date input fields as type="text" where really I wanted type="number"
Also I originally had 3 columns, month and day were each in their own columns, but the lab called for just 2 columns. 
I'm a bit rusty on both SQL and Python since it's been weeks to months since I used either, but with a bit of help from the lecture notes for SQL and w3schools for python I was able to come up with what I needed to get all 4 sections where I wanted them.

I thought about going the extra bit and setting it up so you could at least delete your entry if not delete and update, but I'm really wanting to get started on the final project so I'm going to pass on that for now.