pset9 Finance
Things went pretty well with this one. 
finance.zip has the distributed code.
Had to implement the following pages both the python part in app.py as well as the various .html files: Register, quote/quoted, buy, sell, index, and history. As well as one personal touch. They gave a few suggestions for possibilities. I chose to add a change password page, and the ability to buy or sell from the index page.
 I thought about including keeping track of the last several passwords and limiting the password to at least 8 characters, and none of the previous X passwords re-used, but I had implemented the buy/sell from index and decided that since I did two personal touches I would leave it at just changing the password.

I had wanted to have the buy and sell buttons on the same line as the qty field but things just wouldn't work out for me. I tried messing with the css both with normal css and with bootstrap css stuff and the stuff I could find just wouldn't work so I ended up leaving it like it is.